requirements.txt
bitstream==2.5.4
tqdm==4.50.2
bitstring==3.1.7
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
numpy==1.19.2
Pillow==7.2.0
